user profile and protect route

+++ backend/job/views.py
@@ -1,5 +1,5 @@
 from django.shortcuts import render
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
 from rest_framework import status
 from django.shortcuts import get_object_or_404
@@ -10,6 +10,7 @@ from .serializers import JobSerializer
 from django.db.models import Avg, Min, Max, Count
 from .filters import JobFilters
 from rest_framework.pagination import PageNumberPagination
+from rest_framework.permissions import IsAuthenticated
 
 
 # Create your views here.
@@ -42,7 +43,9 @@ def getJob(request, pk):
 
 
 @api_view(['POST'])
+@permission_classes([IsAuthenticated])
 def newJob(request):
+    request.data['user'] = request.user
     data = request.data
     
     job = Job.objects.create(**data)
@@ -50,8 +53,12 @@ def newJob(request):
     return Response(serializer.data)
 
 @api_view(['PUT'])
+@permission_classes([IsAuthenticated])
 def updateJob(request, pk):
     job = get_object_or_404(Job, id=pk)
+    # check user id based to change
+    if job.user != request.user:
+      return Response({'message': 'you cannoyt update this jobs'}, status=status.HTTP_403_FORBIDDEN)
     
     job.title = request.data['title']
     job.description = request.data['description']
@@ -72,9 +79,13 @@ def updateJob(request, pk):
     return Response(serializer.data)
 
 @api_view(['GET'])
+@permission_classes([IsAuthenticated])
 def deleteJob(request, pk):
     job = get_object_or_404(Job, id=pk)
-
+    
+    if job.user != request.user:
+      return Response({'message': 'you cannot delete this jobs'}, status=status.HTTP_403_FORBIDDEN)
+    
     job.delete()
 
     return Response({ 'message': 'Job is Deleted.' }, status=status.HTTP_200_OK)
